Fix bug about the configuration file
This is a fix for a problem with the configuration file. When autolint
is started for the first time, it will generate a file named
autolint.js. But when for the second time the autolint command is
called, it will try to execute the javascript-file instead of the
command. This is fixed by changing the name of the configuration file so
it doesn't match anymore with the name of the command. Here is the
configuration file called autolintconfig.js.